// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  CONTRACTOR: 'CONTRACTOR',
  FIELD_TEAM: 'FIELD_TEAM'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const PoleStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  MAINTENANCE: 'MAINTENANCE'
} as const

export type PoleStatus = (typeof PoleStatus)[keyof typeof PoleStatus]


export const ReportPriority = {
  P1: 'P1',
  P2: 'P2',
  P3: 'P3'
} as const

export type ReportPriority = (typeof ReportPriority)[keyof typeof ReportPriority]


export const ReportStatus = {
  NEW: 'NEW',
  VERIFIED: 'VERIFIED',
  ASSIGNED: 'ASSIGNED',
  IN_PROGRESS: 'IN_PROGRESS',
  RESOLVED: 'RESOLVED',
  REJECTED: 'REJECTED',
  DUPLICATE: 'DUPLICATE'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const MediaType = {
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO'
} as const

export type MediaType = (typeof MediaType)[keyof typeof MediaType]


export const NotificationChannel = {
  CONSOLE: 'CONSOLE',
  WHATSAPP: 'WHATSAPP',
  SMS: 'SMS',
  EMAIL: 'EMAIL',
  WEB_PUSH: 'WEB_PUSH'
} as const

export type NotificationChannel = (typeof NotificationChannel)[keyof typeof NotificationChannel]


export const NotificationStatus = {
  PENDING: 'PENDING',
  SENT: 'SENT',
  FAILED: 'FAILED'
} as const

export type NotificationStatus = (typeof NotificationStatus)[keyof typeof NotificationStatus]


export const DepartmentType = {
  AUTHORITY: 'AUTHORITY',
  CONTRACTOR: 'CONTRACTOR'
} as const

export type DepartmentType = (typeof DepartmentType)[keyof typeof DepartmentType]